+++ example/src/dev/index.js
@@ -64,3 +64,4 @@ export { default as DatString } from './components/DatString';
 export { default as DatNumber } from './components/DatNumber';
 export { default as DatBoolean } from './components/DatBoolean';
 export { default as DatButton } from './components/DatButton';
+export { default as DatFolder } from './components/DatFolder';